refactor names of ACP projects
