+++ a2a/langgraph/Dockerfile
@@ -1,4 +1,4 @@
-FROM python:3.12-slim
+FROM public.ecr.aws/docker/library/python:3.12-slim
 
 RUN apt-get update && apt-get install -y git
