+++ acp/acp_ollama_researcher/.dockerignore
