+++ acp/acp_ollama_researcher/src/acp_ollama_researcher/prompts.py
