+++ acp/acp_weather_service/src/acp_weather_service/.dockerignore
