+++ acp/acp_weather_service/src/acp_weather_service/graph.py
@@ -4,7 +4,7 @@ from langchain_core.messages import SystemMessage
 from langgraph.prebuilt import tools_condition, ToolNode
 from langchain_openai import ChatOpenAI
 import os
-from ollama_weather_service.configuration import Configuration
+from acp_weather_service.configuration import Configuration
 
 config = Configuration()
